+++ packages/synapse-interface/components/buttons/ButtonLoadingSpinner.tsx
@@ -1,9 +1,0 @@
-import LoadingSpinner from '@tw/LoadingSpinner'
-
-export default function ButtonLoadingSpinner({
-  className,
-}: {
-  className?: string
-}) {
-  return <LoadingSpinner className={`opacity-50 ${className}`} />
-}